organize files and folders by components
